+++ topobank/manager/forms.py
@@ -55,7 +55,10 @@ class TopographyFileUploadForm(forms.ModelForm):
             <a href="{{ cancel_action }}" class="btn btn-default" id="cancel-btn">Cancel</a>
             """),
         ),
-        ASTERISK_HELP_HTML
+        ASTERISK_HELP_HTML,
+        HTML("""
+        For information on the supported file formats, please refer to <a href={% url 'file_formats' %}>this page</a>.
+        """)
     )