+++ topobank/views.py
@@ -7,12 +7,12 @@ from guardian.shortcuts import get_objects_for_user, get_perms_for_model
 
 from allauth.socialaccount.providers.orcid.provider import OrcidProvider
 
-import json
+import markdown2
 
 from termsandconditions.models import TermsAndConditions
+
 from topobank.users.models import User
 from topobank.manager.models import Surface, Topography
-from topobank.manager.utils import selected_instances, selection_choices
 from topobank.analysis.models import Analysis
 
 class HomeView(TemplateView):
@@ -84,5 +84,29 @@ class TermsView(TemplateView):
                 .order_by('optional')
         else:
             context['active_terms'] = active_terms.order_by('optional')
+
+        return context
+
+class FileFormatsView(TemplateView):
+
+    template_name = 'pages/file_formats.html'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data()
+
+        from PyCo.Topography.IO import readers
+
+        reader_infos = []
+        for reader_class in readers:
+            try:
+                # some reader classes have no description yet
+                descr = reader_class.description()
+            except Exception:
+                descr = "*description not yet available*"
+
+            descr = markdown2.markdown(descr)
+
+            reader_infos.append((reader_class.name(), reader_class.format(), descr))
+        context['reader_infos'] = reader_infos
 
         return context